SWIG Updates for Android
The use of BigInteger has been removed in favor of Java long.
AttachCurrentThread is now called in the discovery callback.

Reviewed-on: https://gerrit.iotivity.org/gerrit/27393
Tested-by: IoTivity Jenkins <jenkins-daemon@iotivity.org>

+++ swig/apps/java_lite_simple_client/src/java_lite_simple_client/Client.java
@@ -44,7 +44,7 @@ public class Client {
         }
 
         while (!quit) {
-            long next_event = OCMain.mainPoll().longValue();
+            long next_event = OCMain.mainPoll();
             lock.lock();
             try {
                 if (next_event == 0) {
@@ -57,7 +57,7 @@ public class Client {
                         long next_event_secs = (next_event / CLOCK_TICKS_PER_SECOND);
                         long next_event_nanos = (next_event % CLOCK_TICKS_PER_SECOND) * (NANOS_PER_SECOND / CLOCK_TICKS_PER_SECOND);
 
-                        long now = OCClock.clockTime().longValue();
+                        long now = OCClock.clockTime();
                         long now_secs = (now / CLOCK_TICKS_PER_SECOND);
                         long now_nanos = (now % CLOCK_TICKS_PER_SECOND) * (NANOS_PER_SECOND / CLOCK_TICKS_PER_SECOND);
